refactor: extract read-only chat tool bundles
