+++ python/ai/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Modular ParseChatTools bundles extracted from ai.chat_tools."""
